test(agent): split BYOK encoding boundary coverage
