+++ apps/agent/agent/tests/unit/test_byok_credential_parsing.py
@@ -88,28 +88,6 @@ def test_provider_present_key_missing_rejected() -> None:
     assert excinfo.value.code == "invalid_request"
 
 
-def test_invalid_utf8_key_rejected() -> None:
-    with pytest.raises(ByokError) as excinfo:
-        parse_byok_credential(
-            provider_header="anthropic",
-            key_header=b"\xff",
-            model_header=None,
-            base_url_header=None,
-        )
-    assert excinfo.value.code == "invalid_request"
-
-
-def test_invalid_utf8_base_url_rejected() -> None:
-    with pytest.raises(ByokError) as excinfo:
-        parse_byok_credential(
-            provider_header="openai-compatible",
-            key_header=b"not-a-real-api-key",
-            model_header="test-model",
-            base_url_header=b"https://example.test/\xff",
-        )
-    assert excinfo.value.code == "invalid_request"
-
-
 def test_unknown_provider_rejected() -> None:
     with pytest.raises(ByokError) as excinfo:
         parse_byok_credential(
